cpufreq: Add Relaxed governor
Relaxed is a governor based on chill, but impliments a
boost_ceiling parameter that allows for more gradual boosting.
With chill, I noticed that the min and max frequencies were being
used the most, and all other frequencies were being barely used at
all. Rather than boosting straight to the max frequency, relaxed
finds the differenced between boost_counter and boost_ceiling,
then boosts to max minus that difference. For example, if
boost_count is 8 and boost_ceiling is 12, when boost_counter
reaches 8, the cpu freq will boost to 4 (12-8) levels below max,
then increase if needed untill boost_counter reaches
boost_ceiling. This governor doesn't completely replace chill, but
is intended to be used alongside it. Since the lowest frequency on
a57 cores uses more power than the highest freq on a53 cores, it
is still valuable to use chill on a57 in order to complete a task
as quickly as possible.